tests: Fix LAMBDA-LISTS.SMOKE test
The test invoked the SPECIALIZED-LAMBDA-LIST rule with one argument,
but the rule does not accept any arguments.

The issue went unnoticed because the SPECIALIZED-LAMBDA-LIST rule was
always invoked first in the main test suite and then in the concrete
syntax tree test suite. Both invocations shared the inline cache in
the %RULE-TEST-CASE function. The first invocation caused the inline
cache to be populated with an entry for the SPECIALIZED-LAMBDA-LIST
rule with no arguments. Due to a bug in the inline cache
implementation, the second invocation used this existing cache entry,
thus not passing the (invalid) argument.