definitions: Use EXPRESSION-GRAMMAR:ELEMENT to remove hacks
Despite getting rid of some ugliness, there is an actual improvement:
the previously employed hack used (guard SUB-EXPRESSION (typep 't)) to
force SUB-EXPRESSION to be processed in the element context. However,
as an unwanted side-effect, (typep SUB-EXPRESSION 't) had to be
computed which caused "natural" checks and potential "naturalization"
of SUB-EXPRESSION. This behavior was not wrong per se but placed an
additional burden on the client in terms of implementing and setting
up the expression grammar protocol.